docs(reorg): refactor docs and depandancy to slides
