+++ docs/index.html
@@ -40,7 +40,7 @@
     <script type="module" src="./school-theme/js/sfeir-theme.js"></script>
 
     <link rel="stylesheet" type="text/css" href="./css/pwa.css" id="theme-custo" />
-    <script type="module" src="./scripts/import-slides.js"></script>
+    <script type="module" src="./scripts/reveal-config.js"></script>
   </head>
 
   <body>
@@ -51,6 +51,9 @@
     <script src="./libs/revealjs/head.min.js"></script>
     <script src="./libs/revealjs/reveal.js"></script>
 
-    <script src="reveal-config.js"></script>
+    <!--
+
+      <script src="reveal-config.js"></script>
+    -->
   </body>
 </html>
